requirements.txt
matplotlib==3.8.2
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
scipy==1.11.4
six==1.16.0
tensorflow==2.15.0
flask==3.0.0
flask-cors==4.0.0
protobuf==3.20.3
tensorflow-io==0.30.0